Add test cases for solutions

+++ src/main/kotlin/Solution1.kt
@@ -1,7 +1,7 @@
 import java.io.File
 
 class Solution1 {
-    fun averageNumberOfFieldsPerCsv(files: List<File>): Int {
-        return -1
+    fun averageNumberOfFieldsPerCsv(files: List<File>): Double {
+        return -1.0
     }
 }
